pyproject.toml
[project]
name = "agentbricks-finance"
version = "0.1.0"
description = "AgentBricks Finance Assistant - Multi-agent app with Genie and chat history"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "aiohttp>=3.9.0",
    "openai>=1.52.0",
    "databricks-sdk>=0.40.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
]
